fix: fix IER20LIke decimals

+++ contracts/interfaces/Interfaces.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.7;
 
 interface IERC20Like {
 
-    function decimals() external view returns (uint256 decimals_);
+    function decimals() external view returns (uint8 decimals_);
 
     function balanceOf(address account_) external view returns (uint256 balanceOf_);
